Choreography/consumer/requirements.txt
kafka-python==2.0.2
numpy
scikit-learn
pandas
